Move top level files into micall/ subfolder
