+++ .github/workflows/build-and-test.yml
@@ -119,7 +119,7 @@ jobs:
         run: sudo singularity build micall.simg Singularity
 
       - name: Singularity test
-        run: python release_test_microtest.py micall.simg
+        run: python micall/main.py release_test_microtest micall.simg
 
       - name: Upload Singularity Image
         uses: actions/upload-artifact@v4
