+++ .gitignore
@@ -1,4 +1,4 @@
-/micall_logging_override.py
+/micall/utils/micall_logging_override.py
 /micall/monitor/HIV1_COM_2015_genome_DNA.csv
 /micall/monitor/compare_454_samples.csv
 /micall/monitor/compare_454_samples.png
