+++ micall/main.py
@@ -17,15 +17,6 @@ from importlib.metadata import version
 # Run micall/utils/get_list_of_executables.py to get the up-to-date list of these executables.
 # The consistency of this list is verified in micall/tests/test_installation.py
 EXECUTABLES = [
-    "release_test_publish.py",
-    "micall_kive.py",
-    "micall_watcher.py",
-    "release_test_microtest.py",
-    "docker_build.py",
-    "release_test_setup.py",
-    "release_test_compare.py",
-    "micall_kive_resistance.py",
-    "micall_docker.py",
     "micall/main.py",
     "micall/resistance/genreport.py",
     "micall/resistance/resistance.py",
@@ -48,6 +39,7 @@ EXECUTABLES = [
     "micall/blast_db/make_blast_db.py",
     "micall/utils/concordance_evaluation.py",
     "micall/utils/basespace_upload.py",
+    "micall/utils/release_test_publish.py",
     "micall/utils/compare_mapping.py",
     "micall/utils/project_seeds_from_compendium.py",
     "micall/utils/fasta_to_csv.py",
@@ -61,10 +53,12 @@ EXECUTABLES = [
     "micall/utils/fetch_sequences.py",
     "micall/utils/sam_g2p_simplify.py",
     "micall/utils/contig_summary.py",
+    "micall/utils/micall_kive.py",
     "micall/utils/compare_454_samples.py",
     "micall/utils/genreport_rerun.py",
     "micall/utils/remove_dupe_dirs.py",
     "micall/utils/find_missing_samples.py",
+    "micall/utils/release_test_microtest.py",
     "micall/utils/denovo_simplify.py",
     "micall/utils/sort_sam.py",
     "micall/utils/sample_fastq.py",
@@ -72,21 +66,27 @@ EXECUTABLES = [
     "micall/utils/projects_upload.py",
     "micall/utils/projects_dump.py",
     "micall/utils/find_chimera.py",
+    "micall/utils/docker_build.py",
     "micall/utils/probe_finder.py",
     "micall/utils/aln2counts_simplify.py",
+    "micall/utils/release_test_setup.py",
     "micall/utils/samples_from_454.py",
     "micall/utils/amplicon_finder.py",
     "micall/utils/driver_utils.py",
     "micall/utils/seed_alignments.py",
+    "micall/utils/release_test_compare.py",
     "micall/utils/remap_fastq_simplify.py",
     "micall/utils/contig_counts.py",
+    "micall/utils/micall_kive_resistance.py",
     "micall/utils/ref_aligner.py",
     "micall/utils/scan_run_folders.py",
     "micall/utils/contig_blaster.py",
+    "micall/utils/micall_docker.py",
     "micall/utils/hcv_reference_tree.py",
     "micall/utils/sample_project_summary.py",
     "micall/utils/get_list_of_executables.py",
     "micall/monitor/update_qai.py",
+    "micall/monitor/micall_watcher.py",
     "micall/tcr/igblast.py",
 ]
