+++ micall/tests/test_kive_watcher.py
@@ -23,7 +23,7 @@ from micall.monitor.kive_watcher import find_samples, KiveWatcher, FolderEvent, 
     trim_run_name, compress_old_versions
 from micall.monitor.sample_watcher import PipelineType, ALLOWED_GROUPS, FolderWatcher, SampleWatcher
 from micall.monitor.find_groups import SampleGroup
-from micall_watcher import parse_args
+from micall.monitor.micall_watcher import parse_args
 
 
 class DummyDataset:
